+++ openrouteservice/src/main/java/org/heigit/ors/api/requests/routing/RequestProfileParamsWeightings.java
@@ -27,6 +27,7 @@ public class RequestProfileParamsWeightings {
     public static final String PARAM_STEEPNESS_DIFFICULTY = "steepness_difficulty";
     public static final String PARAM_GREEN_INDEX = "green";
     public static final String PARAM_QUIETNESS = "quiet";
+    public static final String PARAM_SHADOW_INDEX = "shadow";
 
     @ApiModelProperty(name = PARAM_STEEPNESS_DIFFICULTY, value = "Specifies the fitness level for `cycling-*` profiles.\n" +
             "\n level: 0 = Novice, 1 = Moderate, 2 = Amateur, 3 = Pro. The prefered gradient increases with level. CUSTOM_KEYS:{'validWhen':{'ref':'profile','value':['cycling-*']}}", example = "2")
@@ -48,6 +49,13 @@ public class RequestProfileParamsWeightings {
     private Float quietIndex;
     @JsonIgnore
     private boolean hasQuietIndex = false;
+
+    @ApiModelProperty(name = PARAM_SHADOW_INDEX, value = "Specifies the shadow factor for `foot-*` profiles.\n" +
+            "\nfactor: Multiplication factor range from 0 to 1. 0 is the shadow routing base factor without multiplying it by the manual factor and is already different from normal routing. 1 will prefer ways through shadow areas over a shorter route. CUSTOM_KEYS:{'validWhen':{'ref':'profile','value':['foot-*']}}", example = "0.4")
+    @JsonProperty(PARAM_SHADOW_INDEX)
+    private Float shadowIndex;
+    @JsonIgnore
+    private boolean hasShadowIndex = false;
 
     public Integer getSteepnessDifficulty() {
         return steepnessDifficulty;
@@ -76,6 +84,15 @@ public class RequestProfileParamsWeightings {
         hasQuietIndex = true;
     }
 
+    public Float getShadowIndex() {
+        return shadowIndex;
+    }
+
+    public void setShadowIndex(Float shadowIndex) {
+        this.shadowIndex = shadowIndex;
+        hasShadowIndex = true;
+    }
+
     public boolean hasSteepnessDifficulty() {
         return hasSteepnessDifficulty;
     }
@@ -86,5 +103,9 @@ public class RequestProfileParamsWeightings {
 
     public boolean hasQuietIndex() {
         return hasQuietIndex;
+    }
+
+    public boolean hasShadowIndex() {
+        return hasShadowIndex;
     }
 }